Revert "[Chore] #15 - BottomBavBar_Provider 적용"
