+++ ios/Runner/Info.plist
@@ -45,7 +45,5 @@
 		<false/>
 		<key>UIStatusBarHidden</key>
 		<true/>
-		<key>CADisableMinimumFrameDurationOnPhone</key>
-	<true/>
-</dict>
+	</dict>
 </plist>
