+++ lib/src/screens/category.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ttangkkeusmarket/src/widgets/appbar.dart';
+import 'package:ttangkkeusmarket/widgets/appbar.dart';
 
 class CategoryScreen extends StatelessWidget {
   const CategoryScreen({Key? key}) : super(key: key);
