+++ lib/src/screens/home.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:ttangkkeusmarket/src/widgets/appbar.dart';
-import 'package:ttangkkeusmarket/src/widgets/tabbar.dart';
+import 'package:ttangkkeusmarket/widgets/appbar.dart';
+import 'package:ttangkkeusmarket/widgets/tabbar.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
