+++ lib/src/widgets/carousel_list.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:carousel_slider/carousel_slider.dart';
 
 class CarouselSliderWidget extends StatefulWidget {
