+++ pubspec.yaml
@@ -41,7 +41,6 @@ dependencies:
   firebase_auth: ^3.3.17
   cloud_firestore: ^3.1.14
   cloud_firestore_web: ^2.6.14
-  flutter_screenutil: ^5.5.2
 
 dev_dependencies:
   flutter_test:
